+++ src/node/db/API.js
@@ -338,7 +338,7 @@ exports.getChatHistory = async (padID, start, end) => {
   }
 
   // the the whole message-log and return it to the client
-  const messages = await pad.getChatMessages(start, end);
+  const messages = await chat.getChatMessages(pad, start, end);
 
   return {messages};
 };